% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircDataimage.R
\name{CircDataimage}
\alias{CircDataimage}
\title{Generate a graphical user interface (GUI) for interactive circular data images.}
\usage{
CircDataimage()
}
\description{
Generate a graphical user interface (GUI) for interactive circular data images.
}
\examples{
\dontrun{
CircDataimage()
}
}
