% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OceanWind}
\alias{OceanWind}
\title{Ocean Wind Data}
\format{
\code{data.frame()}: \describe{
\item{Vector of time of observation = year + month/12, month in \link{1972.083, 1997.333}}
\item{x,y}{numeric vectors of longitude and latitude}
\item{u,v}{Vector of east and north component of wind (0.01 meters/second).}
}
}
\source{
\url{http://dss.ucar.edu/datasets/ds540.1/data/msga.form.html}
}
\usage{
data('OceanWind')
}
\description{
OceanWind provides a large dataset (495,688 observations) to further explore
the function CircDataimage. The OceanWind data was freely obtained from
ICOADS for the El Nino years 1972, 1976, 1982, 1987, 1991, 1994, and 1997,
January through April, in 1º increments for the area of longitude 0.5º E
to +359.5º E by latitude -59.5º N to +60.5º N.
}
\examples{
data("OceanWind")
head(OceanWind)
}
\keyword{datasets}
