% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{OceanWind}
\alias{OceanWind}
\title{Ocean Wind Data}
\format{
\code{data.frame()}:
\describe{
\item{year}{Vector of time of observation = year + month/12}
\item{x,y}{numeric vectors of longitude and latitude}
\item{u,v}{Vector of east and north component of wind (0.01 meters/second)}
}
}
\source{
\url{http://dss.ucar.edu/datasets/ds540.1/data/msga.form.html}
}
\usage{
data('OceanWind')
}
\description{
OceanWind provides a large dataset (495,688 observations) to further explore
the function \code{\link[=CircDataimage]{CircDataimage()}}. The OceanWind data was freely obtained from
ICOADS for the El Nino years 1972, 1976, 1982, 1987, 1991, 1994, and 1997,
January through April, in 1\eqn{^\circ}{deg} increments for the area of
longitude 0.5\eqn{^\circ}{deg} E to +359.5\eqn{^\circ}{deg} E by latitude
-59.5\eqn{^\circ}{deg} N to +60.5\eqn{^\circ}{deg} N.
}
\examples{
data("OceanWind")
head(OceanWind)
}
\keyword{datasets}
