% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotVectors.R
\name{PlotVectors}
\alias{PlotVectors}
\title{Traditional plots of vector-spatial data.}
\usage{
PlotVectors(
  x,
  y,
  h,
  v,
  UnitVector = TRUE,
  TriIcon = FALSE,
  AdjArrowLength = 1,
  AdjHeadLength = 1,
  TriIconAdj = 1,
  TriRatio = 4,
  JitterPlot = FALSE,
  Jitter = 1,
  ...
)
}
\arguments{
\item{x, y}{numeric. coordinates}

\item{h, v}{numeric vector of horizontal and vertical component.
Missing values are permitted.}

\item{UnitVector}{logical}

\item{TriIcon}{logical}

\item{AdjArrowLength}{Arrow length multiplier}

\item{AdjHeadLength}{Arrow head length multiplier}

\item{TriIconAdj}{Multiplies size of icons}

\item{TriRatio}{Length to width ratio of triangle icon.}

\item{JitterPlot}{logical. If \code{TRUE}, add jitter to location coordinates}

\item{Jitter}{Amount of jitter = Jitter x runif value.}

\item{...}{Additional plot parameters}
}
\description{
Traditional plots of vector-spatial data.
}
\examples{
data(OceanWind)
wind.1997.Jan <- OceanWind[OceanWind$year > 1997 & OceanWind$year < 1997.1, -1]

## Direction Only
PlotVectors(
  x = wind.1997.Jan$x, y = wind.1997.Jan$y, h = wind.1997.Jan$u, v = wind.1997.Jan$v,
  UnitVector = TRUE, AdjArrowLength = 0.75, AdjHeadLength = 0.75, xlim = c(320, 350), ylim = c(0, 30)
)

## Direction and Magnitude
# fields function arrows omits arrowheads with a warning on
# any arrow of length less than 0.001 inch.
PlotVectors(
  x = wind.1997.Jan$x, y = wind.1997.Jan$y, h = wind.1997.Jan$u, v = wind.1997.Jan$v,
  UnitVector = FALSE, TriIcon = FALSE, AdjArrowLength = 3, AdjHeadLength = 0.4, xlim = c(320, 350),
  ylim = c(0, 30)
)

## Triangle Icons
PlotVectors(
  x = wind.1997.Jan$x, y = wind.1997.Jan$y, h = wind.1997.Jan$u, v = wind.1997.Jan$v,
  UnitVector = FALSE, TriIcon = TRUE, TriIconAdj = 0.25, TriRatio = 4, xlim = c(320, 350),
  ylim = c(0, 30)
)
}
