% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCRF.R
\name{SimulateCRF}
\alias{SimulateCRF}
\title{Simulate CRF ~ (Range, CircDistr, Rho, mu=0)}
\usage{
SimulateCRF(
  N = 100,
  CircDistr = c("U", "vM", "WrC", "Tri", "Card"),
  Rho,
  Mu = 0,
  Range,
  Ext = 1,
  CovModel = c("circular", "matern", "exponential", "gaussian", "spherical", "cubic",
    "wave", "power", "powered.exponential", "cauchy", "gencauchy", "gneiting",
    "gneiting.matern", "pure.nugget"),
  xy = NULL,
  Anisotropy = NULL,
  OverFit = FALSE,
  Resolution = 0.01
)
}
\arguments{
\item{N}{Number of spatial locations to simulate}

\item{CircDistr}{Circular distribution in {U, vM, WrC, Tri, Card},}

\item{Rho}{Mean resultant length parameter
\describe{
\item{for triangular}{\verb{0 < Rho <= 4/pi^2}}
\item{for cardioid}{\verb{0 < Rho <= 0.5}}
\item{For vM and wrapped Cauchy}{\verb{0 < Rho < 1}, \verb{1 == }degenerate}
\item{For uniform}{\code{Rho = 0}}
}}

\item{Mu}{argaethsdthsth}

\item{Range}{Distance at which CRV independent}

\item{Ext}{\code{Range*Ext} is horizontal and vertical length of sample space}

\item{CovModel}{Name of spatial correlation function, see package \code{geoR} Help \code{cov.spatial}}

\item{xy}{Regular or irregular \verb{N x 2} matrix of simulation locations, overides \code{N} and \code{Ext}}

\item{Anisotropy}{asrgadh}

\item{Resolution}{drh}

\item{logical.}{\describe{
\item{\code{TRUE}}{standardization (centering and rescaling realization
of the GRV to \code{mean = 0} and \code{sd = 1}) results in closer fit for qualitative evaluation
of the CRV. Undesirable effects are loss of independence of the marginal
GRVs, biased GRF covariance, and biased testing.
Standardization is suitable for demonstration with closer fit,
visualization, and illustrations. Do not standardize for the purposes of
simulation and testing.}
item{\code{FALSE}}{non-standardization (default)
includes expected variation from transformation of variation in mean and
sd of sample of GRV.
}
}}
}
\value{
\describe{
\item{x,y}{Vectors of location coordinates}
\item{direction}{Vector of directions (in radians)}
\item{Z}{Vector of simulated observations of the associated GRV}
}
}
\description{
Simulate CRF ~ (Range, CircDistr, Rho, mu=0)
}
\examples{
xy <- expand.grid(1:11, 1:11) # grid

model.direction1 <- c(
 157, 141, 126, 113, 101, 90, 79, 67, 54, 40, 25, 152, 137, 123, 111,
 100, 90, 80, 69, 57, 44, 30, 147, 133, 120, 109,  99, 90, 81, 71, 60,
 48, 35, 142, 129, 117, 107,  98, 90, 82, 73, 63, 52, 40, 137, 125, 114,
 105,  97, 90, 83, 75, 66, 56, 45, 132, 121, 111, 103,  96, 90, 84, 77,
 69, 60, 50, 127, 117, 108, 101,  95, 90, 85, 79, 72, 64, 55, 122, 113,
 105,  99,  94, 90, 86, 81, 75, 68, 60, 117, 109, 102,  97,  93, 90, 87,
 83, 78, 72, 65, 112, 105,  99,  95,  92, 90, 88, 85, 81, 76, 70, 107,
 101,  96,  93,  91, 90, 89, 87, 84, 80, 75)
 model.direction1 <- (model.direction1)*pi/180

 # Compute vM CRF of 121 observations, Rho=sqrt(0.5) so sill about 0.5,
 from GRF (Range=4, spherical covariance).
 set.seed(666)
 crf1 <- SimulateCRF(CircDistr = "vM", Rho = sqrt(0.5), Range = 4, xy = xy, OverFit = TRUE)
 crf1
}
