% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateCRF.R
\name{SimulateCRF}
\alias{SimulateCRF}
\title{Simulate a circular random field}
\usage{
SimulateCRF(
  N = 100,
  CircDistr = c("vM", "WrC", "Tri", "Card", "U"),
  Rho,
  Mu = 0,
  Range,
  Ext = 1,
  CovModel = c("exponential", "matern", "circular", "gaussian", "spherical", "cubic",
    "wave", "power", "powered.exponential", "cauchy", "gencauchy", "gneiting",
    "gneiting.matern", "pure.nugget"),
  Grid = NULL,
  Anisotropy = NULL,
  OverFit = FALSE,
  Resolution = 0.01
)
}
\arguments{
\item{N}{Number of spatial locations to simulate}

\item{CircDistr}{Circular distribution:
\describe{
\item{"U"}{uniform}
\item{"vM"}{von Mises}
\item{"WrC"}{wrapped Cauchy distributions}
\item{"Tri"}{triangular Cauchy distribution}
\item{"Card}{cardioid Cauchy distributions.}
}}

\item{Rho}{Mean resultant length parameter
\describe{
\item{for triangular}{\verb{0 < Rho <= 4/pi^2}}
\item{for cardioid}{\verb{0 < Rho <= 0.5}}
\item{For vM and wrapped Cauchy}{\verb{0 < Rho < 1}, \verb{1 == }degenerate}
\item{For uniform}{\code{Rho = 0} (all directions have equal density)}
}}

\item{Mu}{Mean resultant direction of circular distribution (rad). \eqn{|\mu| \leq \pi}{mu <= pi}}

\item{Range}{Distance at which CRV are not correlated}

\item{Ext}{\code{Range*Ext} is horizontal and vertical length of sample space}

\item{CovModel}{Name of spatial correlation function, see \code{\link[geoR:cov.spatial]{geoR::cov.spatial()}}}

\item{Grid}{Regular or irregular \verb{N x 2} matrix of simulation locations,
overrides \code{N} and \code{Ext}}

\item{Anisotropy}{Vector of geometric anisotropy (angle, ratio). Angle in
radians, ratio >= 1. See \code{\link[geoR:grf]{geoR::grf()}}}

\item{OverFit}{logical. \describe{
\item{\code{TRUE}}{standardization (centering and rescaling realization
of the GRV to \code{mean = 0} and \code{sd = 1}) results in closer fit for qualitative evaluation
of the CRV. Undesirable effects are loss of independence of the marginal
GRVs, biased GRF covariance, and biased testing.
Standardization is suitable for demonstration with closer fit,
visualization, and illustrations. Do not standardize for the purposes of
simulation and testing.}
\item{\code{FALSE}}{non-standardization (the default)
includes expected variation from transformation of variation in mean and
sd of sample of GRV.
}
}}

\item{Resolution}{For non-closed form inverse CDF, circular quantiles are
interpolated at resolution Resolution. \verb{0.001 <= Resolution <= 0.01}
recommended.}
}
\value{
list with \describe{
\item{x,y}{Vectors of location coordinates}
\item{direction}{Vector of directions (in radians)}
\item{Z}{Vector of simulated observations of the GRV of the GRF.}
}
}
\description{
Simulate a circular random field
}
\examples{
xy <- expand.grid(1:11, 1:11) # grid
SimulateCRF(CircDistr = "vM", Rho = sqrt(0.5), Range = 4, Grid = xy, OverFit = TRUE)
}
