% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestPattern.R
\name{TestPattern}
\alias{TestPattern}
\title{Test Pattern}
\usage{
TestPattern()
}
\value{
list with \describe{
\item{x,y}{Vectors of location coordinates}
\item{u}{Vector of horizontal component of cosines of direction.}
\item{v}{Vector of vertical component of sines of direction.}
}
}
\description{
The function TestPattern makes an intuitive simple
test pattern to explore the function CircDataimage, which produces a GUI for
interactive imaging of circular-spatial data. TestPattern computes direction
such that the direction at any point is the angle between the line from
origin to point and the horizon.
}
\examples{
TestPattern()
}
