% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WorldMask}
\alias{WorldMask}
\title{World Land Mask}
\format{
\code{matrix()} of WorldMask is a matrix of 360 rows (0.5º to 359.5º
longitude) x 121 columns (-59.5º to +60.5º latitude) suited to OceanWind,
with elements NA where wind data is not missing and 1 where wind data is
missing.
}
\usage{
data('WorldMask')
}
\description{
WorldMask is used by \code{\link[=CircDataimage]{CircDataimage()}} to restore land contours to the
circular dataimage of smoothed OceanWind. WorldMask was derived from the R package \code{fields} dataset world.dat.
}
\examples{
data("WorldMask")
head(WorldMask)
}
\keyword{datasets}
