% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\encoding{UTF-8}
\name{WorldMask}
\alias{WorldMask}
\title{World Land Mask}
\format{
\code{matrix()} of WorldMask is a matrix of 360 rows (0.5\eqn{^\circ}{deg}
to 359.5\eqn{^\circ}{deg} longitude) x 121 columns (-59.5\eqn{^\circ}{deg}
to +60.5\eqn{^\circ}{deg} latitude) suited to \code{OceanWind}, with elements \code{NA}
where wind data is not missing and 1 where wind data is missing.
}
\usage{
data('WorldMask')
}
\description{
WorldMask is used by \code{\link[=CircDataimage]{CircDataimage()}} to restore land contours to the
circular dataimage of smoothed \code{OceanWind}. WorldMask was derived from the
R package \code{fields} dataset world.dat.
}
\examples{
data("WorldMask")
head(WorldMask)
}
\keyword{datasets}
